api/requirements.txt
asgiref==3.6.0
Babel==2.12.1
backports.zoneinfo;python_version<"3.9"
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-model-utils==4.3.1
django-money==3.0.0
django-multiselectfield==0.1.12
django-multitenant==3.2.1
django-notifications-hq==1.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.6.0
jsonfield==3.1.0
Markdown==3.4.3
pep8==1.7.1
py-moneyed==2.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.3
swapper==1.3.0
typing-extensions==4.5.0
zipp==3.15.00
gunicorn==20.1.0
django-environ
psycopg2
whitenoise
django-cities-light
django-activity-stream
django-filter
drf-spectacular
django-import-export
django-adminactions
django-simple-history
pandas